parcial3_be/requirements.txt
Django
pymongo==3.12.3
djongo==1.3.6
pytz
dnspython
djangorestframework
django-cors-headers
cloudinary
python-dateutil
djangorestframework-simplejwt
requests
geopy
google-auth